pyproject.toml
[project]
name = "my-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv-python>=0.0.1",
    "google-genai>=1.53.0",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
